app/app/requirements.txt
flask
flask_apscheduler
sqlalchemy
pandas
psycopg2-binary
requests
sklearn